VRVU-38: add new home screen
