+++ app/src/main/java/com/jnj/vaccinetracker/common/di/AndroidModule.kt
@@ -85,6 +85,9 @@ interface AndroidModule {
     fun bindParticipantFlowIntroFragment(): ParticipantFlowIntroFragment
 
     @ContributesAndroidInjector
+    fun bindParticipantFlowAddOrSearchFragment(): ParticipantFlowAddOrSearchFragment
+
+    @ContributesAndroidInjector
     fun bindLoginActivity(): LoginActivity
 
     @ContributesAndroidInjector
